requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
h11==0.14.0
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
undetected-chromedriver==2.1.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
